Bug Fix! X-mode propagation equation is corrected.
Now the variation of the polarization unit vector is captured. In
addition, a keepFFTz parameter is accepted in propagate method, if
given, final result will leave z direction as FFT state, no inverse FFT
is done. This is mainly for ECE2D diagnosis, since FFT in z is required
there.