minor bug fixed: patch touches left boundary
Now, auto_adjust_mesh can correctly treat a situation that one
significant emission area actually extends to the inner boundary of X
coords. This usually means some signal is leaked, so result may be
smaller than it should be. But it will be the user's responsibility to
check the validity of the geometry and parameters.