Choose poloidal harmonics when calculate alpha
Now we can specify a subset of poloidal mode numbers to be included
in the calculated alpha.

New method choose_harmonics accepts an array of integers. Only
positive integers will take effects, and if a number beyond the
read-in mode range is given, the coefficients corresponds to that
will be zero.

method generate_poincare is also updated to accept a array of poloidal mode
numbers. It will call choose_harmonics before calculating the magnetic field
line.

Note that choose_harmonics will change the stored alpha in-place.