BugFix: AlphaDiagnoser FFT convention
The Numpy FFT convention has an additional nagative sign as assumed
before. This causes the formula we used to calculate alpha_mn from GTC
read alpha interpolation coefficients has a mistake in signs on n and m.
This in turn makes our alpha_mn is the complex conjugate of the correct
one.