BUGFIX: float precision induced out of bound error
When evaluating an interpolation on the boundary points, the result
becomes accidentally invalid because of the float number inaccuracy.
A check is added after these evaluations, and if this happens, the
values are re-evaluated at a slightly inner location.

Another physical issue fixed:

When plasma current is not negligibly small, the magnetic field strength
right outside the plasma boundary needs to be artificially decreased
from the simple 1/R formula, so that the resulted B field is still
monotonic. This is required for plasma profile analyzer classes to
search for cold resonant locations.